Rename API module files
